+++ app/javascript/vue/tasks/observation_matrices/matrix_column_coder/store/mutations/mutations.js
@@ -26,6 +26,7 @@ import setSampleStandardMean from './setSampleStandardMean'
 import setSampleUnitFor from './setSampleUnitFor'
 import setPresence from './setPresence'
 import setYearFor from './setYearFor'
+import setDisplayOnlyUnsecored from './options/setDisplayOnlyUnsecored'
 
 const MutationNames = {
   AddObservation: 'addObservation',
@@ -55,7 +56,8 @@ const MutationNames = {
   SetSampleStandardMean: 'setSampleStandardMean',
   SetSampleUnitFor: 'setSampleUnitFor',
   SetPresence: 'setPresence',
-  SetYearFor: 'setYearFor'
+  SetYearFor: 'setYearFor',
+  SetDisplayOnlyUnsecored: 'setDisplayOnlyUnsecored'
 }
 
 const MutationFunctions = {
@@ -86,7 +88,8 @@ const MutationFunctions = {
   [MutationNames.SetSampleStandardMean]: setSampleStandardMean,
   [MutationNames.SetSampleUnitFor]: setSampleUnitFor,
   [MutationNames.SetPresence]: setPresence,
-  [MutationNames.SetYearFor]: setYearFor
+  [MutationNames.SetYearFor]: setYearFor,
+  [MutationNames.SetDisplayOnlyUnsecored]: setDisplayOnlyUnsecored
 }
 
 export {