+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+### Added
+- Add "Display only unsecored rows" on Matrix column coder [#3103]
+
 ### Fixed
 - Presence Descriptor is not saving in Matrix row coder [#3099]
 - Missing number of objects for presence/absence descriptors on Interactive keys [#3102]
@@ -15,6 +18,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 [#3099]: https://github.com/SpeciesFileGroup/taxonworks/issues/3099
 [#3102]: https://github.com/SpeciesFileGroup/taxonworks/issues/3102
+[#3103]: https://github.com/SpeciesFileGroup/taxonworks/issues/3103
 
 ## [0.29.3] - 2022-09-13